requirements.txt
python-binance~=0.7.2
pyyaml~=5.1.1
mysql-connector-python~=8.0.17
